Refactor: deleting deadcode
