+++ drivers/mysql/mysql.go
@@ -93,9 +93,9 @@ func placeholders(numCols, numRows int) string {
 }
 
 // SQLBuilder implements driver.SQLDriver.
-func (d *driveri) SQLBuilder() (*sqlbuilder.Renderer, sqlbuilder.QueryBuilder) {
+func (d *driveri) Renderer() *sqlbuilder.Renderer {
 	r := sqlbuilder.NewDefaultRenderer()
-	return r, &sqlbuilder.BaseQueryBuilder{}
+	return r
 }
 
 // RecordMeta implements driver.SQLDriver.
