+++ drivers/sqlite3/sqlite3.go
@@ -182,9 +182,9 @@ func placeholders(numCols, numRows int) string {
 }
 
 // SQLBuilder implements driver.SQLDriver.
-func (d *driveri) SQLBuilder() (*sqlbuilder.Renderer, sqlbuilder.QueryBuilder) {
+func (d *driveri) Renderer() *sqlbuilder.Renderer {
 	r := sqlbuilder.NewDefaultRenderer()
-	return r, &sqlbuilder.BaseQueryBuilder{}
+	return r
 }
 
 // CopyTable implements driver.SQLDriver.
