+++ libsq/prepare.go
@@ -3,6 +3,8 @@ package libsq
 import (
 	"context"
 
+	"github.com/neilotoole/sq/libsq/ast/sqlbuilder"
+
 	"github.com/neilotoole/sq/libsq/ast"
 	"github.com/neilotoole/sq/libsq/core/errz"
 )
@@ -14,20 +16,18 @@ import (
 // method does this work).
 func (ng *engine) prepare(ctx context.Context, qm *queryModel) error {
 	var (
-		s   string
-		err error
+		err   error
+		frags = &sqlbuilder.Fragments{}
 	)
 
 	// After this switch, ng.bc will be set.
 	switch node := qm.Table.(type) {
 	case *ast.TblSelectorNode:
-		s, ng.targetDB, err = ng.buildTableFromClause(ctx, node)
-		if err != nil {
+		if frags.From, ng.targetDB, err = ng.buildTableFromClause(ctx, node); err != nil {
 			return err
 		}
 	case *ast.JoinNode:
-		s, ng.targetDB, err = ng.buildJoinFromClause(ctx, node)
-		if err != nil {
+		if frags.From, ng.targetDB, err = ng.buildJoinFromClause(ctx, node); err != nil {
 			return err
 		}
 	default:
@@ -35,57 +35,48 @@ func (ng *engine) prepare(ctx context.Context, qm *queryModel) error {
 		return errz.Errorf("unknown selectable %T: %s", node, node)
 	}
 
-	rndr, qb := ng.targetDB.SQLDriver().SQLBuilder()
-	qb.SetFrom(s)
+	rndr := ng.targetDB.SQLDriver().Renderer()
 
-	if s, err = rndr.SelectCols(ng.bc, rndr, qm.Cols); err != nil {
+	if frags.Columns, err = rndr.SelectCols(ng.bc, rndr, qm.Cols); err != nil {
 		return err
 	}
-	qb.SetColumns(s)
 
 	if qm.Distinct != nil {
-		if s, err = rndr.Distinct(ng.bc, rndr, qm.Distinct); err != nil {
+		if frags.Distinct, err = rndr.Distinct(ng.bc, rndr, qm.Distinct); err != nil {
 			return err
 		}
-
-		qb.SetDistinct(s)
 	}
 
 	if qm.Range != nil {
-		if s, err = rndr.Range(ng.bc, rndr, qm.Range); err != nil {
+		if frags.Range, err = rndr.Range(ng.bc, rndr, qm.Range); err != nil {
 			return err
 		}
-
-		qb.SetRange(s)
 	}
 
 	if qm.Where != nil {
-		if s, err = rndr.Where(ng.bc, rndr, qm.Where); err != nil {
+		if frags.Where, err = rndr.Where(ng.bc, rndr, qm.Where); err != nil {
 			return err
 		}
-
-		qb.SetWhere(s)
 	}
 
 	if qm.OrderBy != nil {
-		if s, err = rndr.OrderBy(ng.bc, rndr, qm.OrderBy); err != nil {
+		if frags.OrderBy, err = rndr.OrderBy(ng.bc, rndr, qm.OrderBy); err != nil {
 			return err
 		}
-
-		qb.SetOrderBy(s)
 	}
 
 	if qm.GroupBy != nil {
-		if s, err = rndr.GroupBy(ng.bc, rndr, qm.GroupBy); err != nil {
+		if frags.GroupBy, err = rndr.GroupBy(ng.bc, rndr, qm.GroupBy); err != nil {
 			return err
 		}
-		qb.SetGroupBy(s)
 	}
 
-	ng.targetSQL, err = qb.Render()
-	if err != nil {
-		return err
+	if rndr.PreRender != nil {
+		if err = rndr.PreRender(ng.bc, rndr, frags); err != nil {
+			return err
+		}
 	}
 
-	return nil
+	ng.targetSQL, err = rndr.Render(ng.bc, rndr, frags)
+	return err
 }
